+++ dandy/llm/__init__.py
@@ -1,6 +1,7 @@
 from dandy.llm.bot.llm_bot import BaseLlmBot, LlmBot
 from dandy.llm.prompt.prompt import Prompt
 from dandy.llm.service.config.options import LlmConfigOptions
+from dandy.llm.service.request.message import MessageHistory
 from dandy.llm.map.llm_map import BaseLlmMap
 from dandy.llm.intel import DefaultLlmIntel
 from dandy.map.map import Map
@@ -12,5 +13,6 @@ __all__ = [
     'LlmBot',
     'LlmConfigOptions',
     'Map',
+    'MessageHistory',
     'Prompt',
 ]